% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_chr20_res100kb}
\alias{bulk_chr20_res100kb}
\title{Hi-C data from bulk cell experiments}
\format{
A 625x625 symmetric nonnegative integer-valued matrix.
}
\usage{
bulk_chr20_res100kb
}
\description{
Contact matrix calculated for chromosomes 20 and resolution 100kb (obtained from the Gene Expression Omnibus).
}
\keyword{datasets}
