% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_chr21_res50kb}
\alias{bulk_chr21_res50kb}
\title{Hi-C data from bulk cell experiments}
\format{
A 939x939 symmetric nonnegative integer-valued matrix.
}
\usage{
bulk_chr21_res50kb
}
\description{
Contact matrix calculated for chromosomes 21 and resolution 50kb (obtained from the Gene Expression Omnibus).
}
\keyword{datasets}
