% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_chr1_res100kb_cell8}
\alias{single_chr1_res100kb_cell8}
\title{Hi-C data from single cell experiments}
\format{
A 1924x1924 symmetric nonnegative integer-valued matrix.
}
\usage{
single_chr1_res100kb_cell8
}
\description{
Contact matrix calculated for chromosomes 1, resolution 100kb and cell 8 (obtained from the Gene Expression Omnibus).
}
\keyword{datasets}
