% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPCMS.R
\name{WPCMS}
\alias{WPCMS}
\title{Weighted principal curve metric scaling}
\usage{
WPCMS(
  Z,
  H,
  W = matrix(1, nrow(Z), ncol(Z)),
  Theta = NULL,
  beta = NULL,
  update_beta = TRUE,
  eps = 1e-06,
  maxiter = 100,
  verbose = FALSE
)
}
\arguments{
\item{Z}{a square symmetric matrix.}

\item{H}{a spline basis matrix. By default assumed to have orthogonal columns. If not, orthogonalization should be done via QR decomposition.}

\item{W}{a weight matrix, same dimension as \code{Z}. By default equal weights \code{W = 1} are assumed.}

\item{Theta}{an initialization for spline basis coefficient matrix \eqn{Theta}. If \code{Theta = NULL} the default option \code{Theta = matrix(rnorm(ncol(H) * 3), ncol(H), 3)} is used, i.e. a random initialization is considered.}

\item{beta}{an initialization for intercept \eqn{beta}. If \code{beta = NULL} the default option \code{beta = -min(Z)} is used.}

\item{update_beta}{logical. If \code{update_beta = TRUE}, then the algorithm finds an optimal intercept value. If \code{update_beta = FALSE}, then the intercept is considered to be fixed and set to \code{beta}.}

\item{eps}{a positive convergence tolerance rate. Default value is \code{eps = 1e-6}.}

\item{maxiter}{an integer giving the maximal number of iterations. Default value is \code{maxiter = 100}.}

\item{verbose}{logical. If \code{verbose = TRUE}, the WPCMS loss after each iteration is printed.}
}
\value{
A list containing the WPCMS problem solution:
\itemize{
  \item \code{Theta} -- the matrix of spline parameters.
  \item \code{X} -- the resulting conformation reconstruction.
  \item \code{beta} -- the resulting intercept value.
  \item \code{loss} -- the resulting value of the WPCMS loss.
  \item \code{iter} -- the total number of iterations.
  \item \code{plot} -- the list of WPCMS plots: 'loss' corresponds to loss vs. iteration plot, 'intercept' represents beta vs. iteration plot.
}
}
\description{
WPCMS function calculates the weighted principal curve metric scaling solution for a contact matrix \eqn{Z} (e.g. representing a Hi-C contact matrix after some proper transformation applied) and a spline basis matrix \eqn{H}.
The optimal solution is found via minimizing the weighted Frobenius norm \deqn{ \|W * (Z - D^2(X) + \beta) \|^2_F }{|| W x (Z - D^2(X) + \beta) ||} w.r.t. \eqn{\Theta} subject to the smooth curve constraint \eqn{X = H\Theta}.
Here \eqn{D(X)} refers to the matrix of pairwise distances.
The spatial coordinates of the resulting reconstruction are presented in \eqn{X}.
}
\examples{
data(C)

#transform contact counts to distances
Z = 1/(C+1)

#create spline basis matrix
H = splines::bs(1:ncol(C), df = 5)

#orthogonalize H using QR decomposition
H = qr.Q(qr(H))

#run WPCMS with equal weights; optimize intercept 
WPCMS(Z, H)$X

#run WPCMS with random weights; fixed intercept
W = matrix(runif(length(Z)), nrow(Z), ncol(Z))
WPCMS(Z, H, W, beta = 1, update_beta = FALSE)$X

}
